feat: Ability to have a specially named executable to install a certain modpack and version (or latest)
fixes #18, fixes #19

+++ CurseForge.Minecraft.Serverpack.Launcher/Classes/CurseForgeFile.cs
@@ -5,9 +5,9 @@ namespace CurseForge.Minecraft.Serverpack.Launcher
 	public class CurseForgeFile
 	{
 		[JsonPropertyName("projectID")]
-		public long ProjectId { get; set; }
+		public uint ProjectId { get; set; }
 		[JsonPropertyName("fileID")]
-		public long FileId { get; set; }
+		public uint FileId { get; set; }
 		[JsonPropertyName("required")]
 		public bool Required { get; set; }
 	}
